Add extra reducer to load data from asyncstorage
