+++ screens/IntroScreen/IntroScreen.js
@@ -4,7 +4,7 @@ import { View, StyleSheet, Dimensions, Animated, StatusBar } from 'react-native'
 import { useDispatch } from 'react-redux';
 import { COLORS, icons } from '../../constants';
 import { setInstalled } from '../../redux/appReducer';
-import { resetAllCat } from '../../redux/catReducer';
+import { loadCategoriesFromStorage, resetAllCat } from '../../redux/catReducer';
 //Redux
 import { Slide, SubSlide, Ticker, Pagination } from './components';
 import slides from './IntroSlides';
@@ -16,109 +16,6 @@ export const IntroScreen = () => {
   const scrollX = useRef(new Animated.Value(0)).current;
   const scrollClick = useRef(null);
   const unmounted = useRef(false);
-  const income = 'income';
-  const expense = 'expense';
-
-  const cat = [
-    {
-      id: 1,
-      name: 'Vente',
-      icon: icons.shopping,
-      cat: income,
-      color: COLORS.purple,
-      data: [],
-    },
-    {
-      id: 2,
-      name: 'Remboursement',
-      icon: icons.refund,
-      cat: income,
-      color: COLORS.blue,
-      data: [],
-    },
-    {
-      id: 3,
-      name: 'Intérêt',
-      icon: icons.interest,
-      cat: income,
-      color: COLORS.darkgreen,
-      data: [],
-    },
-    {
-      id: 4,
-      name: 'Subvention',
-      icon: icons.grant,
-      cat: income,
-      color: COLORS.red,
-      data: [],
-    },
-    {
-      id: 5,
-      name: 'Investissement',
-      icon: icons.investment,
-      cat: income,
-      color: COLORS.peach,
-      data: [],
-    },
-
-    {
-      id: 6,
-      name: 'Achat',
-      icon: icons.shopping,
-      cat: expense,
-      color: COLORS.lightBlue,
-      data: [],
-    },
-    {
-      id: 7,
-      name: 'Salaire',
-      icon: icons.cash,
-      cat: expense,
-      color: COLORS.peach,
-      data: [],
-    },
-    {
-      id: 8,
-      name: "Dépenses d'exploitation",
-      icon: icons.cashbook,
-      cat: expense,
-      color: COLORS.darkgreen,
-      data: [],
-    },
-    {
-      id: 9,
-      name: "Retraits d'argent",
-      icon: icons.sell,
-      cat: expense,
-      color: COLORS.red,
-      data: [],
-    },
-    {
-      id: 10,
-      name: 'Paiements de dettes',
-      icon: icons.income,
-      cat: expense,
-      color: COLORS.yellow,
-      data: [],
-    },
-    {
-      id: 11,
-      name: 'Autres entrées',
-      icon: icons.more,
-      cat: income,
-      color: COLORS.gray,
-      data: [],
-    },
-
-    {
-      id: 12,
-      name: 'Autres Sorties',
-      icon: icons.more,
-      cat: expense,
-      color: COLORS.purple,
-      data: [],
-    },
-  ];
 
   useEffect(() => {
     return () => {
@@ -144,11 +41,14 @@ export const IntroScreen = () => {
       .catch((error) => console.log('Error: ', error));
 
     // async storage add default cats and dispath it
-    await AsyncStorage.setItem('categories', JSON.stringify(cat))
-      .then(async (json) => {
-        dispatch(resetAllCat(cat));
-      })
-      .catch((error) => console.log('Error: ', error));
+
+    dispatch(loadCategoriesFromStorage());
+
+    // await AsyncStorage.setItem('categories', JSON.stringify(cat))
+    //   .then(async (json) => {
+    //     dispatch(resetAllCat(cat));
+    //   })
+    //   .catch((error) => console.log('Error: ', error));
   };
 
   return (
